darc-arc: lift the extraction loop into the library
Groundwork for porting `Unarc/`, and useful on its own.

`run_blocks` walks the data blocks, decrypts, decompresses, verifies each file's
CRC and either reports or writes it. It lived in `bin/darc.rs`, so it was
reachable only by the `darc` binary -- and `unarc` and every SFX module need
exactly that work. It moves to `extract.rs`, beside `Layout` and `is_safe`,
which it already used. `show3`/`ratio3` come with it, being its output format.

**One implementation, two callers.** That is the point rather than a
convenience: a second copy of this logic is precisely how `Unarc/ArcStructure.h`
came to read the per-file time field as 4 bytes while the writer wrote 8, so
directories extracted as empty files and every CRC failed while the listing
looked perfect. The C++ reader is being retired rather than ported alongside
this one, so nothing will diverge from it again.

Pure refactor -- no behaviour change, and the gates say so:

  696 tests
  arc-golden-check   118 cases, 0 differing
  arc-extract-check  24 extractions, 0 differing
  arc-cli-check      24 cases, 0 differing
  clippy gate clean

Two notes for whoever repeats this kind of move. Deleting the function by line
range also ate the closing brace of `ratio3`, which the compiler caught; and the
`sed -e 's/\barchive::/...'` that was supposed to rewrite the module paths did
nothing at all, because BSD sed has no `\b`. Both were caught by building, but
the second is the kind that passes silently when the pattern happens not to
matter.
